LayoutTests/http/tests/security/XFrameOptions/x-frame-options-multiple-headers-sameorigin-allow.html

<!DOCTYPE html>
<html>
<head>
    <script>
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.dumpChildFramesAsText();
            testRunner.dumpResourceLoadCallbacks();
        }
    </script>
</head>
<body>
    <p>The frame below should load, proving that 'sameorigin, sameorigin' === 'sameorigin'.</p>
    <iframe style="width:500px; height:500px" src="http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-multiple-headers-sameorigin.cgi"></iframe>
</body>
</html>
